Refactor: organize benchmarks into proper namespaces/files
All benchmark classes now use the explicit PixelConverterBenchmarks.Benchmarks namespace. Restored all necessary using directives and class definitions. Each benchmark class resides in its own file with correct namespace and usings. No logic changes; this is an organizational refactor for improved structure and clarity.
