+++ benchmarks/SIPSorceryMedia.Abstractions/PixelConverterBenchmarks/Benchmarks/ConvertNv12ToI420Benchmarks.cs
@@ -2,17 +2,17 @@
 using BenchmarkDotNet.Attributes;
 using SIPSorceryMedia.Abstractions;
 
-namespace PixelConverterBenchmarks;
+namespace PixelConverterBenchmarks.Benchmarks;
 
-public class ConvertKnownI420ToBGRBenchmarks
+public class ConvertNv12ToI420Benchmarks
 {
 #if !LibVersion
-    CommunityToolkit.HighPerformance.Buffers.ArrayPoolBufferWriter<byte>? _bgrWriter = new();
+    CommunityToolkit.HighPerformance.Buffers.ArrayPoolBufferWriter<byte>? _i420Writer = new();
 #endif
 
     public IEnumerable<BenchmarkParams> GetImages()
     {
-        yield return CreateBenchmarkParams("ref-i420.yuv", 640, 480);
+        yield return CreateBenchmarkParams("ref-nv12.yuv", 640, 480);
 
         static BenchmarkParams CreateBenchmarkParams(string image, int width, int height)
         {
@@ -28,27 +28,27 @@ public class ConvertKnownI420ToBGRBenchmarks
     public void IterationSetup()
     {
 #if !LibVersion
-        Debug.Assert(_bgrWriter is not null);
-        _bgrWriter.Clear();
+        Debug.Assert(_i420Writer is not null);
+        _i420Writer.Clear();
 #endif
     }
 
     [Benchmark]
-    public int ConvertKnownI420ToBGRArray()
+    public int ConvertNV12ToI420Array()
     {
-        var bgr = PixelConverter.I420toBGR(Image.Bytes, Image.Width, Image.Height, out _);
-        return bgr.Length;
+        var i420 = PixelConverter.NV12toI420(Image.Bytes, Image.Width, Image.Height);
+        return i420.Length;
     }
 
     [Benchmark]
-    public int ConvertKnownI420ToBGRBufferWriter()
+    public int ConvertNV12ToI420BufferWriter()
     {
 #if LibVersion
         return 0;
 #else
-        Debug.Assert(_bgrWriter is not null);
-        PixelConverter.I420toBGR(_bgrWriter, Image.Bytes, Image.Width, Image.Height, out _);
-        return _bgrWriter.WrittenCount;
+        Debug.Assert(_i420Writer is not null);
+        PixelConverter.NV12toI420(_i420Writer, Image.Bytes.AsSpan(), Image.Width, Image.Height);
+        return _i420Writer.WrittenCount;
 #endif
     }
 }
